chore: pages/api 폴더 제거 후 /service/api 폴더 구조 변경
