--- pages/api/axiosInstance.ts
@@ -1,16 +1,0 @@
-import axios from "axios";
-
-const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
-
-const axiosInstance = axios.create({
-  baseURL: BASE_URL,
-  timeout: 5000,
-  headers: {
-    "Access-Control-Allow-Origin": "*",
-    "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, OPTIONS",
-    "Access-Control-Allow-Headers":
-      "Origin, X-Requested-With, Content-Type, Accept",
-  },
-});
-
-export default axiosInstance;
