+++ pages/boards/[id]/components/Comment/index.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import kebabIcon from "@/assets/icons/ic_kebab.svg";
 import profileImage from "@/assets/icons/ic_profile.svg";
 import styles from "./styles.module.scss";
-import { Article } from "@/pages/api/articles";
+import { Article } from "@/services/articles";
 
 interface CommentProps {
   comment: Article;
