+++ pages/boards/components/All/AllPost/index.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import styles from "./styles.module.scss";
 import heartIcon from "@/assets/icons/ic_heart.svg";
 import profileImage from "@/assets/icons/ic_profile.svg";
-import { Article } from "@/pages/api/articles";
+import { Article } from "@/services/articles";
 import Link from "next/link";
 
 interface AllPostProps {
