+++ pages/boards/components/Best/BestArticleItem/index.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import styles from "./styles.module.scss";
 import heartIcon from "@/assets/icons/ic_heart.svg";
-import { Article } from "@/pages/api/articles";
+import { Article } from "@/services/articles";
 import bestBadge from "@/assets/images/img_badge.svg";
 import Link from "next/link";
 interface BestArticleItemProps {
