+++ services/axiosInstance.ts
@@ -1,0 +1,10 @@
+import axios from "axios";
+
+const BASE_URL = process.env.NEXT_PUBLIC_API_URL;
+
+const axiosInstance = axios.create({
+  baseURL: BASE_URL,
+  timeout: 5000,
+});
+
+export default axiosInstance;
